Convert textDelta to ptrdiff_t

+++ frontend/mipsdecoder.h
@@ -21,8 +21,8 @@ class MIPSDecoder : public NJMCDecoder {
 public:
 	MIPSDecoder(Prog *prog);
 
-	virtual DecodeResult &decodeInstruction(ADDRESS pc, int delta);
-	virtual int decodeAssemblyInstruction(ADDRESS pc, int delta);
+	virtual DecodeResult &decodeInstruction(ADDRESS pc, ptrdiff_t delta);
+	virtual int decodeAssemblyInstruction(ADDRESS pc, ptrdiff_t delta);
 
 private:
 	/*
