+++ frontend/pentiumdecoder.h
@@ -23,8 +23,8 @@ class PentiumDecoder : public NJMCDecoder {
 public:
 	PentiumDecoder(Prog *prog);
 
-	virtual DecodeResult &decodeInstruction(ADDRESS pc, int delta);
-	virtual int decodeAssemblyInstruction(ADDRESS pc, int delta);
+	virtual DecodeResult &decodeInstruction(ADDRESS pc, ptrdiff_t delta);
+	virtual int decodeAssemblyInstruction(ADDRESS pc, ptrdiff_t delta);
 
 private:
 	/*
