+++ frontend/st20decoder.h
@@ -23,8 +23,8 @@ class ST20Decoder : public NJMCDecoder {
 public:
 	ST20Decoder(Prog *prog);
 
-	virtual DecodeResult &decodeInstruction(ADDRESS pc, int delta);
-	virtual int decodeAssemblyInstruction(ADDRESS pc, int delta);
+	virtual DecodeResult &decodeInstruction(ADDRESS pc, ptrdiff_t delta);
+	virtual int decodeAssemblyInstruction(ADDRESS pc, ptrdiff_t delta);
 
 private:
 	/*
